refactor: Split select_basic.rs into smaller test files
- Split 737-line select_basic.rs into three focused files:
  - select_basic_projection.rs: Basic projection tests (SELECT *, specific columns)
  - select_order_by.rs: ORDER BY clause tests
  - select_derived_columns.rs: SQL:1999 derived column lists tests
- Updated mod.rs to include new test modules
- Removed original select_basic.rs file

This improves maintainability and makes tests easier to find by functionality.

Amp-Thread-ID: https://ampcode.com/threads/T-135e46fb-64ba-46e4-b12b-092ce7c4da21
